+++ server/mongo/middleware/auth.js
@@ -1,4 +1,4 @@
-const { User } = require('../models/User');
+const { User } = require('../schemas/User');
 
 let auth = (req, res, next) => {
   let token = req.cookies.w_auth;
