+++ server/mongo/routes/logs/login.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const router = express.Router();
-const { User } = require("../../models/User");
+const { User } = require("../../schemas/User");
 
 router.post("/", (req, res) => {
     User.findOne({ email: req.body.email }, (err, user) => {
