+++ server/mongo/routes/logs/logout.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const router = express.Router();
-const { User } = require("../../models/User");
+const { User } = require("../../schemas/User");
 
 router.get("/logout", (req, res) => {
   User.findOneAndUpdate({ _id: req.user._id }, { token: "", tokenExp: "" }, (err, doc) => {
